more binder

+++ src/NfcGonkMessage.h
@@ -87,12 +87,12 @@ typedef enum {
  * @see NFCForum-TS-NDEF, clause 3.2
  */
 typedef struct {
-  uint32_t tnf;
+  uint8_t tnf;
 
   uint32_t typeLength;
   uint8_t* type;
 
-  uint32_t idLength;
+  uint8_t idLength;
   uint8_t* id;
 
   //TODO : For Short Record, length is only 1 byte.
@@ -161,12 +161,12 @@ typedef struct {
   /**
    * The NDEF is read-only or not.
    */
-  uint32_t isReadOnly;
+  uint8_t isReadOnly;
 
   /**
    * The NDEF can be configured to read-only or not.
    */
-  uint32_t canBeMadeReadonly;
+  uint8_t canBeMadeReadonly;
 
   /**
    * Maximum length of the NDEF.
